-> Changed components location into a new folder
