+++ StockHut/StockHut/ClientApp/src/components/Home.js
@@ -3,31 +3,49 @@ import logo from "../login/stockhut.png";
 import "../login/login.css";
 import { Route } from "react-router"
 import { NavLink, Link } from "react-router-dom";
-import  SignUp  from "../login/signup.js";
+import LogIn from "./LogIn"
 
-const Home = () => {
-   
-    return (
-        
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <div
-          className="login-page"
-          style={{ backgroundColor: "rgb(199, 255, 171)" }}
-        >
-          <div className="form">
-            <form className="login-form">
-              <input type="text" placeholder="username" />
-              <input type="password" placeholder="password" />
-            <button>login</button>
-            <NavLink tag={Link} to="/signup">Create an account</NavLink>
-            </form>
-          </div>
-        </div>
-      </header>
-    </div>
-  );
+class Home extends Component {
+
+
+constructor(props) {
+    super(props);
+    this.state = {
+        username: [],
+        password: [],
+        user: "",
+        pass: ""
+    }
+
+    this.handleChange = this.handleChange.bind(this)
+}
+    
+componentDidMount() {
+
+    fetch('https://localhost:44314/api/SignIn')
+        .then(response => response.json())
+        .then(data => {
+            this.setState(prevState => {
+                username: data.map((obj) => { prevState.username.push(obj.username) })
+                password: data.map((obj) => { prevState.password.push(obj.password) })
+            })
+        })
+    }
+
+    handleChange(event) {
+        this.setState({
+            [event.target.name]: event.target.value
+        })
+    }
+
+    render() {
+        return (
+            <div>
+                <LogIn handleChange={this.handleChange} user={this.state.user} pass={this.state.pass} />
+                <h1>{this.state.user} {this.state.pass} {console.log(this.state.user)} {console.log(this.state.pass)}</h1>
+            </div>
+        );
+    }
 };
 
 export default Home;
