+++ StockHut/StockHut/ClientApp/src/components/Navbar/NavMenu.js
@@ -1,0 +1,31 @@
+﻿import React, { Component } from "react";
+import "./NavMenu.css";
+
+import ButtonAppBar from "./Navbar";
+
+export class NavMenu extends Component {
+    static displayName = NavMenu.name;
+
+    constructor(props) {
+        super(props);
+
+        this.toggleNavbar = this.toggleNavbar.bind(this);
+        this.state = {
+            collapsed: true
+        };
+    }
+
+    toggleNavbar() {
+        this.setState({
+            collapsed: !this.state.collapsed
+        });
+    }
+
+    render() {
+        return (
+            <header>
+                <ButtonAppBar />
+            </header>
+        );
+    }
+}
