+++ StockHut/StockHut/ClientApp/src/components/Navbar/Navbar.js
@@ -1,0 +1,37 @@
+﻿import React from "react";
+import { makeStyles } from "@material-ui/core/styles";
+import AppBar from "@material-ui/core/AppBar";
+import Toolbar from "@material-ui/core/Toolbar";
+import Typography from "@material-ui/core/Typography";
+import Button from "@material-ui/core/Button";
+import IconButton from "@material-ui/core/IconButton";
+import MenuIcon from "@material-ui/icons/Menu";
+
+import TemporaryDrawer from "./NavDrawer";
+
+const useStyles = makeStyles(theme => ({
+    root: {
+        flexGrow: 1
+    },
+    menuButton: {
+        marginRight: theme.spacing(2)
+    },
+    title: {
+        flexGrow: 1
+    },
+    login: {
+        color: 'white'
+    }
+}));
+
+function ButtonAppBar() {
+    const classes = useStyles();
+
+    return (
+        <div className={classes.root}>
+           
+        </div>
+    );
+}
+
+export default ButtonAppBar;
