--- StockHut/StockHut/ClientApp/src/components/Profile.js
@@ -1,1 +1,0 @@
-﻿
