+++ StockHut/StockHut/StockHut.csproj
@@ -20,9 +20,21 @@
 
   <ItemGroup>
     <!-- Don't publish the SPA source files, but do show them in the project files list -->
+    <Compile Remove="ClientApp\src\components\NewFolder1\**" />
+    <Compile Remove="ClientApp\src\components\NewFolder2\**" />
     <Content Remove="$(SpaRoot)**" />
+    <Content Remove="ClientApp\src\components\NewFolder1\**" />
+    <Content Remove="ClientApp\src\components\NewFolder2\**" />
+    <EmbeddedResource Remove="ClientApp\src\components\NewFolder1\**" />
+    <EmbeddedResource Remove="ClientApp\src\components\NewFolder2\**" />
     <None Remove="$(SpaRoot)**" />
+    <None Remove="ClientApp\src\components\NewFolder1\**" />
+    <None Remove="ClientApp\src\components\NewFolder2\**" />
     <None Include="$(SpaRoot)**" Exclude="$(SpaRoot)node_modules\**" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="ClientApp\src\components\FeedPage\" />
   </ItemGroup>
 
   <Target Name="DebugEnsureNodeEnv" BeforeTargets="Build" Condition=" '$(Configuration)' == 'Debug' And !Exists('$(SpaRoot)node_modules') ">
